agent/requirements.txt
# Lehigh Mars Rover AI Agent - Separate Environment
fastapi
uvicorn
python-dotenv
pytest
pinecone
openai
PyMuPDF
youtube-transcript-api>=0.6.3
